requirements.txt
psycopg2>=2.8.3
python-telegram-bot>=12.1.1
pyzmq>=18.1.0
